16-feb/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
google-generativeai==0.8.5
google-cloud-firestore==2.19.0
google-cloud-texttospeech==2.21.1
google-cloud-translate==3.16.0
google-cloud-storage==2.18.2
google-cloud-aiplatform==1.72.0
pydantic==2.11.1
